requirements.txt
pyyaml==5.1
opencv-python
numpy
pandas
